projects/hotel-realtime-pipeline/requirements.txt
kafka-python-ng>=2.2.0
pandas>=2.0.0
duckdb>=1.0.0
pyarrow>=14.0.0
faker>=20.0.0
python-dotenv>=1.0.0
dbt-duckdb>=1.7.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
watchdog>=4.0.0
